Remove workflow
This action didn't work well because of the strangeness of win32com.client.EnsureDispatch. I'm sure there's a way to get it to
build, but for now let's retire this and use it as an example of how to
commit and push if we ever need to self-host another runner on Windows.